refactor: update AppInstallationParameters to use contentfulApiKey and adjust related components for improved API key handling
