+++ apps/iterable/test/mocks/locations/Sidebar.spec.tsx
@@ -1,7 +1,8 @@
-import Sidebar from './Sidebar';
+import React from 'react';
 import { render } from '@testing-library/react';
-import { mockCma, mockSdk } from '../../test/mocks';
 import { vi } from 'vitest';
+import { mockSdk, mockCma } from '..';
+import Sidebar from '../../../src/locations/Sidebar';
 
 vi.mock('@contentful/react-apps-toolkit', () => ({
   useSDK: () => mockSdk,
